Initial implementation of ceilometer-alarm
The implemented use case is raising an alarm (i.e. executing something)
as soon as a cumulative counter is above or below a certain threshold.

The result of the "statistic" during a "period" is compared
to the alarm "threshold" with the operator in "comparaison_operator"
field of the alarm.

The criteria handled to check if the alarm threshold needs to be checked
for a meter are for now only:
- counter_name
- resource_id

When a alarm is triggered a URL is called, same things when the
alarm is untriggered

The URLs are stored in fields "ok_action" and "alarm_action" of the
alarm description.

A api permit to made CRUD operation on the alarm.
This one update the DB and send a rpc message to ceilometer-alarm to
inform it for alarm creation/update/deletion.

This code is latest one for this prototype.
The summit have defined a more convenient way to acheive the alerting.

Change-Id: I2dfebc30a47576e83906a6d4011ce5d52041e901